refactor(agw): remove duplicated implementations of DataModel methods (#13564)
