+++ lte/gateway/python/magma/enodebd/devices/baicells_qrtb.py
@@ -11,7 +11,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 from distutils.util import strtobool
-from typing import Any, Callable, Dict, List, Optional
+from typing import Any, Callable, Dict
 
 from dp.protos.cbsd_pb2 import CBSDStateResult
 from magma.common.service import MagmaService
@@ -906,93 +906,13 @@ class BaicellsQRTBTrDataModel(DataModel):
             is_optional=False,
         )
 
+    TRANSFORMS_FOR_ENB: Dict[ParameterName, Callable[[Any], Any]] = {}
     TRANSFORMS_FOR_MAGMA = {
         # We don't set GPS, so we don't need transform for enb
         ParameterName.GPS_LAT: transform_for_magma.gps_tr181,
         ParameterName.GPS_LONG: transform_for_magma.gps_tr181,
     }
     PARAMETERS.update(CarrierAggregationParameters.CA_PARAMETERS)  # noqa: WPS604
-
-    @classmethod
-    def get_parameter(cls, param_name: ParameterName) -> Optional[TrParam]:
-        """
-        Retrieve parameter by its name
-
-        Args:
-            param_name (ParameterName): parameter name to retrieve
-
-        Returns:
-            Optional[TrParam]
-        """
-        return cls.PARAMETERS.get(param_name)
-
-    @classmethod
-    def _get_magma_transforms(
-            cls,
-    ) -> Dict[ParameterName, Callable[[Any], Any]]:
-        return cls.TRANSFORMS_FOR_MAGMA
-
-    @classmethod
-    def _get_enb_transforms(cls) -> Dict[ParameterName, Callable[[Any], Any]]:
-        return {}
-
-    @classmethod
-    def get_load_parameters(cls) -> List[ParameterName]:
-        """
-        Retrieve all load parameters
-
-        Returns:
-             List[ParameterName]
-        """
-        return cls.LOAD_PARAMETERS
-
-    @classmethod
-    def get_num_plmns(cls) -> int:
-        """
-        Retrieve the number of all PLMN parameters
-
-        Returns:
-            int
-        """
-        return cls.NUM_PLMNS_IN_CONFIG
-
-    @classmethod
-    def get_parameter_names(cls) -> List[ParameterName]:
-        """
-        Retrieve all parameter names
-
-        Returns:
-            List[ParameterName]
-        """
-        excluded_params = [
-            str(ParameterName.DEVICE),
-            str(ParameterName.FAP_SERVICE),
-        ]
-        names = list(
-            filter(
-                lambda x: (not str(x).startswith('PLMN')) and (str(x) not in excluded_params),
-                cls.PARAMETERS.keys(),
-            ),
-        )
-        return names
-
-    @classmethod
-    def get_numbered_param_names(cls) -> Dict[ParameterName, List[ParameterName]]:
-        """
-        Retrieve parameter names of all objects
-
-        Returns:
-            Dict[ParameterName, List[ParameterName]]
-        """
-        names = {}
-        for i in range(1, cls.NUM_PLMNS_IN_CONFIG + 1):
-            params = []
-            params.append(ParameterName.PLMN_N_CELL_RESERVED % i)
-            params.append(ParameterName.PLMN_N_ENABLE % i)
-            params.append(ParameterName.PLMN_N_PRIMARY % i)
-            params.append(ParameterName.PLMN_N_PLMNID % i)
-            names[ParameterName.PLMN_N % i] = params
-        return names
 
 
 class BaicellsQRTBTrConfigurationInitializer(EnodebConfigurationPostProcessor):
