+++ lte/gateway/python/magma/enodebd/devices/experimental/cavium.py
@@ -11,7 +11,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
-from typing import Any, Callable, Dict, List, Optional, Type
+from typing import Any, Dict, Optional, Type
 
 from magma.common.service import MagmaService
 from magma.enodebd.data_models import transform_for_enb, transform_for_magma
@@ -302,6 +302,8 @@ class CaviumTrDataModel(DataModel):
     Class to represent relevant data model parameters from TR-196/TR-098/TR-181.
     This class is effectively read-only
     """
+    # Parameters to query when reading eNodeB config
+    LOAD_PARAMETERS = [ParameterName.DEVICE]
     # Mapping of TR parameter paths to aliases
     DEVICE_PATH = 'Device.'
     FAPSERVICE_PATH = DEVICE_PATH + 'Services.FAPService.1.'
@@ -429,60 +431,6 @@ class CaviumTrDataModel(DataModel):
         ParameterName.GPS_LAT: transform_for_magma.gps_tr181,
         ParameterName.GPS_LONG: transform_for_magma.gps_tr181,
     }
-
-    @classmethod
-    def get_parameter(cls, param_name: ParameterName) -> Optional[TrParam]:
-        return cls.PARAMETERS.get(param_name)
-
-    @classmethod
-    def _get_magma_transforms(
-        cls,
-    ) -> Dict[ParameterName, Callable[[Any], Any]]:
-        return cls.TRANSFORMS_FOR_MAGMA
-
-    @classmethod
-    def _get_enb_transforms(cls) -> Dict[ParameterName, Callable[[Any], Any]]:
-        return cls.TRANSFORMS_FOR_ENB
-
-    @classmethod
-    def get_load_parameters(cls) -> List[ParameterName]:
-        """
-        Load all the parameters instead of a subset.
-        """
-        return [ParameterName.DEVICE]
-
-    @classmethod
-    def get_num_plmns(cls) -> int:
-        return cls.NUM_PLMNS_IN_CONFIG
-
-    @classmethod
-    def get_parameter_names(cls) -> List[ParameterName]:
-        excluded_params = [
-            str(ParameterName.DEVICE),
-            str(ParameterName.FAP_SERVICE),
-        ]
-        names = list(
-            filter(
-                lambda x: (not str(x).startswith('PLMN'))
-                and (str(x) not in excluded_params),
-                cls.PARAMETERS.keys(),
-            ),
-        )
-        return names
-
-    @classmethod
-    def get_numbered_param_names(
-        cls,
-    ) -> Dict[ParameterName, List[ParameterName]]:
-        names = {}
-        for i in range(1, cls.NUM_PLMNS_IN_CONFIG + 1):
-            params = []
-            params.append(ParameterName.PLMN_N_CELL_RESERVED % i)
-            params.append(ParameterName.PLMN_N_ENABLE % i)
-            params.append(ParameterName.PLMN_N_PRIMARY % i)
-            params.append(ParameterName.PLMN_N_PLMNID % i)
-            names[ParameterName.PLMN_N % i] = params
-        return names
 
 
 class CaviumTrConfigurationInitializer(EnodebConfigurationPostProcessor):
