+++ lte/gateway/python/magma/enodebd/devices/freedomfi_one.py
@@ -11,7 +11,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 import logging
-from typing import Any, Callable, Dict, Iterable, List, Optional
+from typing import Any, Callable, Dict, Iterable, List
 
 from dp.protos.cbsd_pb2 import CBSDStateResult
 from magma.enodebd.data_models import transform_for_magma
@@ -911,7 +911,7 @@ class FreedomFiOneTrDataModel(DataModel):
             is_optional=False,
         ),
     }
-    TRANSFORMS_FOR_ENB: Dict[ParameterName, Callable[[Any], Any]] = {}
+
     NUM_PLMNS_IN_CONFIG = 1
     for i in range(1, NUM_PLMNS_IN_CONFIG + 1):  # noqa: WPS604
         PARAMETERS[ParameterName.PLMN_N % i] = TrParam(
@@ -953,6 +953,10 @@ class FreedomFiOneTrDataModel(DataModel):
     PARAMETERS.update(StatusParameters.DERIVED_STATUS_PARAMETERS)  # noqa: WPS604
     PARAMETERS.update(CarrierAggregationParameters.CA_PARAMETERS)  # noqa: WPS604
 
+    # Parameters to query when reading eNodeB config
+    LOAD_PARAMETERS = list(PARAMETERS.keys())
+
+    TRANSFORMS_FOR_ENB: Dict[ParameterName, Callable[[Any], Any]] = {}
     TRANSFORMS_FOR_MAGMA = {
         # We don't set these parameters
         ParameterName.BAND_CAPABILITY: transform_for_magma.band_capability,
@@ -960,92 +964,6 @@ class FreedomFiOneTrDataModel(DataModel):
         ParameterName.GPS_LAT: transform_for_magma.gps_tr181,
         ParameterName.GPS_LONG: transform_for_magma.gps_tr181,
     }
-
-    @classmethod
-    def get_parameter(cls, param_name: ParameterName) -> Optional[TrParam]:
-        """
-        Retrieve parameter by its name
-
-        Args:
-            param_name (ParameterName): parameter name to retrieve
-
-        Returns:
-            Optional[TrParam]
-        """
-        return cls.PARAMETERS.get(param_name)
-
-    @classmethod
-    def _get_magma_transforms(
-            cls,
-    ) -> Dict[ParameterName, Callable[[Any], Any]]:
-        return cls.TRANSFORMS_FOR_MAGMA
-
-    @classmethod
-    def _get_enb_transforms(cls) -> Dict[ParameterName, Callable[[Any], Any]]:
-        return cls.TRANSFORMS_FOR_ENB
-
-    @classmethod
-    def get_load_parameters(cls) -> List[ParameterName]:
-        """
-        Retrieve all load parameters
-
-        Returns:
-             List[ParameterName]
-        """
-        return list(cls.PARAMETERS.keys())
-
-    @classmethod
-    def get_num_plmns(cls) -> int:
-        """
-        Retrieve the number of all PLMN parameters
-
-        Returns:
-            int
-        """
-        return cls.NUM_PLMNS_IN_CONFIG
-
-    @classmethod
-    def get_parameter_names(cls) -> List[ParameterName]:
-        """
-        Retrieve all parameter names
-
-        Returns:
-            List[ParameterName]
-        """
-        excluded_params = [
-            str(ParameterName.DEVICE),
-            str(ParameterName.FAP_SERVICE),
-        ]
-        names = list(
-            filter(
-                lambda x: (not str(x).startswith('PLMN'))
-                and (str(x) not in excluded_params),
-                cls.PARAMETERS.keys(),
-            ),
-        )
-        return names
-
-    @classmethod
-    def get_numbered_param_names(
-            cls,
-    ) -> Dict[ParameterName, List[ParameterName]]:
-        """
-        Retrieve parameter names of all objects
-
-        Returns:
-            Dict[ParameterName, List[ParameterName]]
-        """
-        names = {}
-        for i in range(1, cls.NUM_PLMNS_IN_CONFIG + 1):
-            params = [
-                ParameterName.PLMN_N_CELL_RESERVED % i,
-                ParameterName.PLMN_N_ENABLE % i,
-                ParameterName.PLMN_N_PRIMARY % i,
-                ParameterName.PLMN_N_PLMNID % i,
-            ]
-            names[ParameterName.PLMN_N % i] = params
-
-        return names
 
     @classmethod
     def get_sas_param_names(cls) -> Iterable[ParameterName]:
